requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.5
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pypdf==6.9.2
pillow==12.1.1
psycopg2-binary==2.9.11
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2025.3
gunicorn==23.0.0
